+++ messaging/coap/coap.h
@@ -246,10 +246,10 @@ int coap_get_header_uri_host(
   const char **host); /* in-place string might not be 0-terminated. */
 int coap_set_header_uri_host(void *packet, const char *host);
 
-int coap_get_header_uri_path(
+size_t coap_get_header_uri_path(
   void *packet,
   const char **path); /* in-place string might not be 0-terminated. */
-int coap_set_header_uri_path(void *packet, const char *path, int path_len);
+size_t coap_set_header_uri_path(void *packet, const char *path, size_t path_len);
 
 int coap_get_header_uri_query(
   void *packet,
